Add loop detection for intersections

+++ src/render/texture.rs
@@ -29,14 +29,14 @@ impl Texture {
         Self { image }
     }
 
-    pub fn flood_fill(&mut self, x: u32, y: u32, color: Rgba<u8>, wrap_x: bool, wrap_y: bool) {
-        let start_color = self.image.get_pixel(x, y);
-        self.image.put_pixel(x, y, color);
+    pub fn flood_fill(&mut self, x: i32, y: i32, color: Rgba<u8>, wrap_x: bool, wrap_y: bool) {
+        let start_color = self.image.get_pixel(x as u32, y as u32);
+        self.image.put_pixel(x as u32, y as u32, color);
         let mut to_color = vec![(x, y)];
 
         while let Some((x, y)) = to_color.pop() {
-            if self.image.get_pixel(x, y) == start_color {
-                self.image.put_pixel(x, y, color);
+            if self.image.get_pixel(x as u32, y as u32) == start_color {
+                self.image.put_pixel(x as u32, y as u32, color);
 
                 self.add_for_fill(&mut to_color, x + 1, y, start_color, wrap_x, wrap_y);
                 self.add_for_fill(&mut to_color, x, y + 1, start_color, wrap_x, wrap_y);
@@ -48,26 +48,26 @@ impl Texture {
 
     fn add_for_fill(
         &self,
-        to_color: &mut Vec<(u32, u32)>,
-        mut x: u32,
-        mut y: u32,
+        to_color: &mut Vec<(i32, i32)>,
+        mut x: i32,
+        mut y: i32,
         start_color: Rgba<u8>,
         wrap_x: bool,
         wrap_y: bool,
     ) {
         if wrap_x {
-            x %= self.image.width();
+            x %= self.image.width() as i32;
         }
 
         if wrap_y {
-            y %= self.image.height();
+            y %= self.image.height() as i32;
         }
 
         if x > 0
             && y > 0
-            && x < self.image.width()
-            && y < self.image.height()
-            && self.image.get_pixel(x, y) == start_color
+            && (x as u32) < self.image.width()
+            && (y as u32) < self.image.height()
+            && self.image.get_pixel(x as u32, y as u32) == start_color
         {
             to_color.push((x, y));
         }
